+++ modules/apps/forms-and-workflow/calendar/calendar-test/src/testIntegration/java/com/liferay/calendar/test/util/CalendarBookingTestUtil.java
@@ -372,6 +372,20 @@ public class CalendarBookingTestUtil {
 			null, 0, null, serviceContext);
 	}
 
+	public static CalendarBooking addRegularCalendarBookingWithReminders(
+			Calendar calendar, long startTime, long endTime, int firstReminder,
+			int secondReminder)
+		throws PortalException {
+
+		User user = UserLocalServiceUtil.getUser(calendar.getUserId());
+
+		return addCalendarBooking(
+			user, calendar, new long[0], RandomTestUtil.randomLocaleStringMap(),
+			RandomTestUtil.randomLocaleStringMap(), startTime, endTime, null,
+			firstReminder, NotificationType.EMAIL, secondReminder,
+			NotificationType.EMAIL, createServiceContext(user));
+	}
+
 	public static CalendarBooking
 			addRegularCalendarBookingWithTitleAndDescription(
 				User user, Calendar calendar, Map<Locale, String> titleMap,